# contract/openapi.json
{
  "components": {
    "schemas": {
      "ActivityCategory": {
        "description": "The ONLY valid category values anywhere in the system (CONTRACTS.md\n\u00a72, \u00a78) \u2014 budget-breakdown grouping, seed data, and frontend filters\nall depend on this exact set. Add a value here first if it's ever\nneeded, never ad hoc in one layer.",
        "enum": [
          "sightseeing",
          "food",
          "adventure",
          "transport",
          "stay",
          "other"
        ],
        "title": "ActivityCategory",
        "type": "string"
      },
...
      "ActivityRead": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ActivityCategory"
          },
          "city_id": {
            "format": "uuid",
            "title": "City Id",
            "type": "string"
          },
          "cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Hours"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "city_id",
          "category"
        ],
        "title": "ActivityRead",
        "type": "object"
      },
...
      "CityRead": {
        "properties": {
          "cost_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Index"
          },
          "country": {
            "title": "Country",
            "type": "string"
          },
          "google_place_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Google Place Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "popularity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Popularity"
          }
        },
        "required": [
          "id",
          "name",
          "country",
          "latitude",
          "longitude"
        ],
        "title": "CityRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ItineraryActivityCreate": {
        "description": "POST /stops/{id}/activities (CONTRACTS \u00a74).\n\n`cost` is optional on the way in: when omitted the route copies\n`Activity.cost` at insert time (CONTRACTS \u00a72). Passing it explicitly\nlets a user record what they actually expect to pay for this trip.",
        "properties": {
          "activity_id": {
            "format": "uuid",
            "title": "Activity Id",
            "type": "string"
          },
          "cost": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "scheduled_date": {
            "format": "date",
            "title": "Scheduled Date",
            "type": "string"
          },
          "scheduled_time": {
            "anyOf": [
              {
                "exclusiveMaximum": 24.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Time"
          }
        },
        "required": [
          "activity_id",
          "scheduled_date"
        ],
        "title": "ItineraryActivityCreate",
        "type": "object"
      },
      "ItineraryActivityRead": {
        "properties": {
          "activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "activity_id": {
            "format": "uuid",
            "title": "Activity Id",
            "type": "string"
          },
          "cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "scheduled_date": {
            "format": "date",
            "title": "Scheduled Date",
            "type": "string"
          },
          "scheduled_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Time"
          },
          "stop_id": {
            "format": "uuid",
            "title": "Stop Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stop_id",
          "activity_id",
          "scheduled_date"
        ],
        "title": "ItineraryActivityRead",
        "type": "object"
      },
...
      "StopCreate": {
        "description": "POST /trips/{id}/stops.\n\n`sequence` is optional: omit it and the route appends the stop to the\nend of the trip, which is what the Itinerary Builder does for every\n\"add city\" click. The cross-row date rules (CONTRACTS \u00a72) are checked\nin services/stops.py at write time, not here.",
        "properties": {
          "city_id": {
            "format": "uuid",
            "title": "City Id",
            "type": "string"
          },
          "date_end": {
            "format": "date",
            "title": "Date End",
            "type": "string"
          },
          "date_start": {
            "format": "date",
            "title": "Date Start",
            "type": "string"
          },
          "sequence": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence"
          }
        },
        "required": [
          "city_id",
          "date_start",
          "date_end"
        ],
        "title": "StopCreate",
        "type": "object"
      },
      "StopRead": {
        "description": "Stop as returned by the API.\n\nThe last five fields are computed per request by services/feasibility.py\n(CONTRACTS \u00a77.2) \u2014 they are NOT columns. The frontend displays them and\nnever recomputes any of them itself (CONTRACTS \u00a78).",
        "properties": {
          "city": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CityRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "city_id": {
            "format": "uuid",
            "title": "City Id",
            "type": "string"
          },
          "date_end": {
            "format": "date",
            "title": "Date End",
            "type": "string"
          },
          "date_start": {
            "format": "date",
            "title": "Date Start",
            "type": "string"
          },
          "distance_from_previous_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance From Previous Km"
          },
          "distance_source": {
            "anyOf": [
              {
                "enum": [
                  "directions",
                  "haversine"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Source"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_feasible": {
            "default": true,
            "title": "Is Feasible",
            "type": "boolean"
          },
          "itinerary_activities": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ItineraryActivityRead"
            },
            "title": "Itinerary Activities",
            "type": "array"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "travel_duration_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Duration Hours"
          },
          "travel_gap_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Gap Days"
          },
          "trip_id": {
            "format": "uuid",
            "title": "Trip Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "trip_id",
          "city_id",
          "sequence",
          "date_start",
          "date_end"
        ],
        "title": "StopRead",
        "type": "object"
      },
      "StopReorderItem": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sequence": {
            "minimum": 0.0,
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "sequence"
        ],
        "title": "StopReorderItem",
        "type": "object"
      },
      "StopUpdate": {
        "properties": {
          "city_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Id"
          },
          "date_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date End"
          },
          "date_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Start"
          },
          "sequence": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence"
          }
        },
        "title": "StopUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "GlobeTrotter API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/stops/reorder": {
      "patch": {
        "description": "Apply a whole new ordering in ONE request (CONTRACTS \u00a74: batched, never\none call per row) \u2014 a drag-reorder in the builder is a single write.\n\nEvery stop must belong to the same trip, and that trip must be one the\ncaller may touch. Both are checked before anything is mutated, so a\npartially-authorized batch changes nothing at all.",
        "operationId": "reorder_stops_stops_reorder_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/StopReorderItem"
                },
                "title": "Payload",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StopRead"
                  },
                  "title": "Response Reorder Stops Stops Reorder Patch",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Reorder Stops",
        "tags": [
          "stops"
        ]
      }
    },
    "/stops/{stop_id}": {
      "delete": {
        "operationId": "delete_stop_stops__stop_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Stop",
        "tags": [
          "stops"
        ]
      },
      "patch": {
        "operationId": "update_stop_stops__stop_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stop Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StopUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Stop",
        "tags": [
          "stops"
        ]
      }
    },
    "/stops/{stop_id}/activities": {
      "get": {
        "operationId": "list_stop_activities_stops__stop_id__activities_get",
        "parameters": [
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ItineraryActivityRead"
                  },
                  "title": "Response List Stop Activities Stops  Stop Id  Activities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Stop Activities",
        "tags": [
          "stops"
        ]
      },
      "post": {
        "description": "Schedule a catalog activity on this stop.\n\n`cost` defaults from `Activity.cost` at insert time and is independent\nfrom then on (CONTRACTS \u00a72), so re-pricing the catalog later never\nrewrites a planned trip's numbers.",
        "operationId": "add_stop_activity_stops__stop_id__activities_post",
        "parameters": [
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stop Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItineraryActivityCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryActivityRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add Stop Activity",
        "tags": [
          "stops"
        ]
      }
    },
...
    "/trips/{trip_id}/stops": {
      "get": {
        "operationId": "list_stops_trips__trip_id__stops_get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StopRead"
                  },
                  "title": "Response List Stops Trips  Trip Id  Stops Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Stops",
        "tags": [
          "stops"
        ]
      },
      "post": {
        "description": "Add a city leg. Both cross-row date rules run before the insert\n(services/stops.py) rather than being re-derived here.",
        "operationId": "create_stop_trips__trip_id__stops_post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StopCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Stop",
        "tags": [
          "stops"
        ]
      }
    },
...
  }
}